Move `create_duckdb`
